+++ CHANGELOG.md
@@ -27,6 +27,10 @@ Pyxel doesn't use SemVer anymore, since most minor releases have at least minor 
   model groups.
   (See [!1080](https://gitlab.com/esa/pyxel/-/merge_requests/1080)
   and [!1081](https://gitlab.com/esa/pyxel/-/merge_requests/1081)).
+* Add more catalogs (e.g. Hipparcos, Tycho2) for model 
+  [`load_star_map`](https://esa.gitlab.io/pyxel/doc/stable/references/model_groups/scene_generation_models.html#load-star-map)
+  from [Scene Generation](https://esa.gitlab.io/pyxel/doc/stable/references/model_groups/scene_generation_models.html#).
+  (See [!1083](https://gitlab.com/esa/pyxel/-/merge_requests/1083)).
 
 ### Others
 * Add compatibility with `lacosmic` 1.3+.